refactor(organizations): map logo to first table column
