+++ backend/src/modules/organization/controllers/organization.controller.ts
@@ -34,7 +34,7 @@ import { OrganizationFilterDto } from '../dto/organization-filter.dto';
 import { UpdateOrganizationDto } from '../dto/update-organization.dto';
 import { Organization } from '../entities';
 import { OrganizationRequest } from '../entities/organization-request.entity';
-import { OrganizationView } from '../entities/organization.view-entity';
+import { OrganizationView } from '../entities/organization-view.entity';
 import { OrganizationRequestService } from '../services/organization-request.service';
 import { OrganizationService } from '../services/organization.service';
