+++ backend/src/modules/organization/entities/index.ts
@@ -1,11 +1,11 @@
-export * from './organization-general.entity';
-export * from './organization-activity.entity';
-export * from './organization-legal.entity';
-export * from './organization-financial.entity';
-export * from './organization-report.entity';
-export * from './organization.entity';
 export * from './contact.entity';
-export * from './report.entity';
-export * from './partner.entity';
 export * from './investor.entity';
-export * from './organization.view-entity';
+export * from './organization-activity.entity';
+export * from './organization-financial.entity';
+export * from './organization-general.entity';
+export * from './organization-legal.entity';
+export * from './organization-report.entity';
+export * from './organization-view.entity';
+export * from './organization.entity';
+export * from './partner.entity';
+export * from './report.entity';
